requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.2.0
langsmith>=0.2.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# OpenAI integration
openai>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
deepeval>=0.2.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
aiofiles>=23.2.0
python-multipart>=0.0.6

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
